fix(contracts): payerKind width + §4.5 nonce correction + ADR 0013
Follow-ups to the merged marketplace contract surface (#627), per operator
review of the three open decisions.

§4.5 — operator confirmed the spec prose was wrong, not the impl. EIP-3009
burns the nonce inside transferWithAuthorization; checking `revoked` before
that call means a revoked-mid-flight buy reverts cleanly with the nonce
UNSPENT. Replay risk is zero because revocation is monotonic. Corrected the
spec prose and added ADR 0013 pinning the monotonic-revocation invariant.
Implementation unchanged (was already correct).

payerKind — widened the `Sold.payerKind` event field uint8 -> uint16 (headroom
for richer payer identity: EOA / smart-account / intent-solver / sponsored)
while purchasePath stays a tight uint8 enum. Marked it a v1 placeholder
("mirrors purchasePath; do NOT derive analytics") in the Solidity event
NatSpec, the Rust sol! binding, and the subgraph schema. Logged the derivation
refinement ticket in the program plan §7.1 deferred-items register.

evm_version (shanghai bump) is handled separately — gated on a 5003 smoke
deploy per the operator.

Verified: forge build, forge test (58 passed / 1 fork skipped), forge fmt
--check, cargo fmt --check + cargo check -p xvision-identity -p
xvision-marketplace.

+++ contracts/src/interfaces/IMarketplace.sol
@@ -31,6 +31,13 @@ interface IMarketplace {
 
     function feeRecipient() external view returns (address);
 
+    /// @dev `payerKind` is a **v1 placeholder** — in v1 it merely mirrors
+    ///      `purchasePath` (0/1). DO NOT derive analytics from it yet. It is
+    ///      typed `uint16` (wider than `purchasePath`'s tight enum) so the
+    ///      eventual derivation can encode richer payer identity (EOA vs smart
+    ///      account vs intent solver vs sponsored, possibly composite) without
+    ///      an ABI change. Refinement ticket tracked in the marketplace program
+    ///      plan §7.1 deferred-items register.
     event Sold(
         uint256 indexed listingId,
         uint256 indexed agentNftId,
@@ -39,7 +46,7 @@ interface IMarketplace {
         uint96 sellerProceeds,
         uint96 protocolProceeds,
         uint256 licenseTokenId,
-        uint8 payerKind, // 0 = human, 1 = agent
+        uint16 payerKind, // v1 placeholder (mirrors purchasePath); see note above
         uint8 purchasePath // 0 = direct, 1 = x402
     );
     event ProtocolFeeBpsChanged(uint16 oldBps, uint16 newBps);
